Refactor input handling by removing unused hooks and integrating their logic directly into the Home component
